backend/requirements.txt
python-dotenv
langchain-community
pypdf
pinecone
fastapi[standard]
groq
pymongo[srv]